Moved Linux-specific runtime-manager code to it's own crate
